Enhance README with additional TODO items and links
Updated the README to include new tasks such as adding a link to the GitHub repository and the RSS feed, as well as suggestions for using more emojis. Also improved the web page enhancement tasks by specifying the use of a CSS framework and adding a search feature to the HTML interface.